Moved postponed report controller
